Work in progress on collisions
