+++ index.html
@@ -18,7 +18,6 @@
       </p>
       <div id="game"></div>
     </div>
-    <script type="text/javascript" src="dist/phaser.js"></script>
     <script type="text/javascript">
       window.onload = function() {
         var script = document.createElement('script');
